Validate Add/Edit/Delete Board/List/Video

+++ src/containers/BoardRead.jsx
@@ -25,7 +25,7 @@ class BoardRead extends React.Component {
 
   constructor(props) {
     super(props)
-    this.state = { isEditing: false, editingBoardPart: { title: '', slug: '', error: null } }
+    this.state = { isEditing: false, title: '', slug: '', error: null }
     this.onTitleClick = this.onTitleClick.bind(this)
     this.onInputBlur = this.onInputBlur.bind(this)
     this.onInputChange = this.onInputChange.bind(this)
@@ -40,10 +40,7 @@ class BoardRead extends React.Component {
     const title = currentBoard.title
     const slug = currentBoard.slug
 
-    this.setState({
-      isEditing: true,
-      editingBoardPart: { ...this.state.editingBoardPart, title, slug }
-    })
+    this.setState({ isEditing: true, title, slug })
 
     // TODO: focus() on <input>
   }
@@ -55,22 +52,27 @@ class BoardRead extends React.Component {
   onInputChange(event) {
     const title = event.target.value
     const slug = title.trim().toString().toLowerCase().replace(/\s+/g, '-')
-    this.setState({
-      editingBoardPart: {
-        ...this.state.editingBoardPart,
-        title,
-        slug,
-        error: slug === 'trash' && 'Reserved board title'
-      }
-    })
+    const boardExists = _.find(
+      this.props.videoStorage.boards,
+      board => {return slug === board.slug && slug !== this.props.params.boardSlug}
+    )
+    let error = null
+
+    if (slug === 'trash') {
+      error = 'Reserved board title'
+    } else if (boardExists) {
+      error = 'Board already exists'
+    }
+
+    this.setState({ title, slug, error })
   }
 
   onPressEnter() {
     const currentBoard = _.find(this.props.videoStorage.boards, board => {
       return board.slug === this.props.params.boardSlug
     })
-    const title = this.state.editingBoardPart.title.trim()
-    const { slug, error } = this.state.editingBoardPart
+    const title = this.state.title.trim()
+    const { slug, error } = this.state
 
     if (title && slug && !error) {
       this.props.editBoard(currentBoard, { title, slug })
@@ -84,7 +86,7 @@ class BoardRead extends React.Component {
       return board.slug === this.props.params.boardSlug
     })
 
-    if (confirm(`Delete ${currentBoard.title}?`)) {
+    if (confirm(`Delete ${currentBoard.title}?\nAll lists and videos will be deleted.`)) {
       this.props.deleteBoard(currentBoard)
       this.context.router.push('/')
     }
@@ -114,10 +116,13 @@ class BoardRead extends React.Component {
               onBlur={this.onInputBlur}
               onChange={this.onInputChange}
               onKeyPress={event => {if (event.key === 'Enter') this.onPressEnter()}}
-              value={this.state.editingBoardPart.title}
+              value={this.state.title}
               ref={input => {this.boardTitleInput = input}}
             />
-            <small>{this.state.editingBoardPart.error}</small>
+
+            {this.state.error &&
+              <small>{this.state.error}</small>
+            }
           </header>
         }
 
@@ -127,7 +132,7 @@ class BoardRead extends React.Component {
               _.find(
                 this.props.videoStorage.videos,
                 video => {
-                  return video.board === currentBoard.title && !video.list && !video.deleted
+                  return video.board === currentBoard.slug && !video.list && !video.deleted
                 }
               ) &&
               <div className="VideoWrapper">
@@ -153,7 +158,7 @@ class BoardRead extends React.Component {
             <div className="VideoWrapper">
               <article className="Card">
                 <header className="CardHeader">
-                  <ListAdd boardSlug={currentBoard.slug} />
+                  <ListAdd board={currentBoard} />
                 </header>
               </article>
             </div>
